+++ MonitorLabelsPlugin.cs
@@ -10,7 +10,9 @@ namespace MonitorLabels
 	{
 		public const string GUID = "DannyVD.mods.LethalCompany.MonitorLabels";
 		public const string PLUGIN_NAME = "MonitorLabels";
-		public const string PLUGIN_VERSION = "1.1.0"; //TODO: Consider making this 2.0.0
+		public const string PLUGIN_VERSION = "1.2.0"; //TODO: Consider making this 2.0.0
+		
+		public const string DEPENDENCY_STRING = $"DannyVD-{PLUGIN_NAME}-{PLUGIN_VERSION}";
 
 		private void Awake()
 		{
@@ -20,7 +22,7 @@ namespace MonitorLabels
 			LoggerUtil.Initialize(ConfigUtil.LoggingLevel, Logger);
 
 			// Plugin startup logic
-			LoggerUtil.Log(LogLevel.Info, $"Plugin {GUID} is loaded!"); // Using the Log function circumvents the configuration option, this is by design
+			LoggerUtil.Log(LogLevel.Info, $"Plugin {DEPENDENCY_STRING} is loaded!"); // Using the Log function circumvents the configuration option, this is by design
 
 			PatchUtil.PatchFunctions();
 		}